move `doc.rs` to its own subdirectory
